+++ CHANGELOG.md
@@ -16,6 +16,36 @@ All notable changes to this project are documented in this file.
 
 - n/a
 
+## [0.3.0] - tbd
+
+### Added
+
+- Create and use user-specific Salt configuration.
+- Check for "sudo" being installed before using it.
+- Add "initmaster" command.
+- Use host-specific ssh keys for salt-ssh.
+- Extend usage info.
+
+### Changed
+
+- Only configure FUSE "allow_other" if required
+
+### Fixed
+
+- Add further reason to read config again
+- Correctly return result of get_os_id()
+
+## [0.2.0] - 2025-02-15
+
+### Added
+
+- Add "initremote" command.
+- Add "startshell" command.
+
+### Changed
+
+- Only install Salt config if it is not already valid.
+
 ## [0.1.0] - 2024-11-24
 
 ### Added